ENH: Added function for determinnig the set of live registers at any given instruction in the DFG.  ENH: Reordered the addition of kernel prefix variables so that it will work even if prepending code for uninitialized values.

+++ ocelot/executive/interface/LLVMExecutableKernel.h
@@ -73,7 +73,7 @@ namespace executive
 			
 		private:
 			/*! \brief Determine the padding required to satisfy alignment */
-			unsigned int _pad( unsigned int& size, unsigned int alignment );
+			unsigned int _pad( size_t& size, unsigned int alignment );
 		
 		private:
 			/*! \brief Create the LLVM module from the code */
